Rename package policyv2 to policy
