+++ cyral/internal/policy/datasource.go
@@ -1,4 +1,4 @@
-package policyv2
+package policy
 
 import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
