+++ cyral/internal/policy/model.go
@@ -1,4 +1,4 @@
-package policyv2
+package policy
 
 import (
 	"context"
